requirements.txt
mcrcon
python-dotenv
py-dactyl